Moved command related files to sub folder in bloom
