+++ src/bloom/commands/BloomCommandInvoker.cpp
