+++ src/bloom/commands/IBloomCommand.h
