+++ src/utils/BloomInputProccesor.cpp
@@ -4,9 +4,9 @@
 #include <vector>
 #include "data_persistence/IDataPersistence.h"
 #include "data_persistence/FilePersistence.h"
-#include "bloom/InsertUrlCommand.h"
-#include "bloom/CheckUrlCommand.h"
-#include "bloom/BloomCommandInvoker.h"
+#include "bloom/commands/InsertUrlCommand.h"
+#include "bloom/commands/CheckUrlCommand.h"
+#include "bloom/commands/BloomCommandInvoker.h"
 #include "utils/InputValidation.h"
 
 /**
